Improve how vnic detachment is handled when deleting an instance.
When deleting an instance, don't try to explicitly delete vncis attached to that instance. Trying to delete a prinary vnic results in an error that we used to ignore and print confusing messages to debug log. Ignoring errors was necessary because there is no API available to determine whether a vnic is primary or secondary. Delete instance behavior is that all vnics attached to the instance being deleted will be detached *eventually*.  Eager/explict vnic detachment was an attempt to reduce the time it takes for the vnic detachment to complete (so that the same IP can be resued for a different instance).

Now vnic detachment is handled as
 - Issue a delete instance request
 - Wait for all attached vnics to be detached
 - Wait for instance to be deleted